// app/dashboard/jobs/job-type.ts
export interface JobType {
    id: string,
    occupation: string;
    employmentType: string;
    date: string;
    deadline: string;
    company: string;
    location: string;
    summary: string;
    contractType: string;
    contractTypeOther: string;
    salary: string;
    workingHours: string;
    holidays: string;
    age: string;
    jobNumber: string;
    publicRange: string;
    jobLabels: string[];
    jobCount: number;
}